Removed default state from filter mod cycle

+++ Configs.cs
@@ -62,8 +62,8 @@ namespace tMusicPlayer
 				if (ResettingPanels) {
 					musicPlayerUI.MusicPlayerPanel.Left.Pixels = 500f;
 					musicPlayerUI.MusicPlayerPanel.Top.Pixels = 6f;
-					musicPlayerUI.selectionPanel.Left.Pixels = (float)(Main.screenWidth / 2) - musicPlayerUI.selectionPanel.Width.Pixels / 2f;
-					musicPlayerUI.selectionPanel.Top.Pixels = (float)(Main.screenHeight / 2) - musicPlayerUI.selectionPanel.Height.Pixels / 2f;
+					musicPlayerUI.selectionPanel.Left.Pixels = (Main.screenWidth / 2) - musicPlayerUI.selectionPanel.Width.Pixels / 2f;
+					musicPlayerUI.selectionPanel.Top.Pixels = (Main.screenHeight / 2) - musicPlayerUI.selectionPanel.Height.Pixels / 2f;
 					musicPlayerUI.musicEntryPanel.Left.Pixels = musicPlayerUI.selectionPanel.Left.Pixels + musicPlayerUI.selectionPanel.Width.Pixels - musicPlayerUI.musicEntryPanel.Width.Pixels - 10f;
 					musicPlayerUI.musicEntryPanel.Top.Pixels = musicPlayerUI.selectionPanel.Top.Pixels - musicPlayerUI.musicEntryPanel.Height.Pixels;
 					ResettingPanels = false;
